requirements.txt
Appium-Python-Client
pytest
pytest-allure-adaptor